clean source code
