--- src/main/java/Game/ControlHangmanScene.java
@@ -1,4 +1,0 @@
-package Game;
-
-public class ControlHangmanScene {
-}
