+++ src/main/java/Game/Game.java
@@ -11,7 +11,7 @@ public class Game extends Application {
 
     @Override
     public void start(Stage stage) throws IOException {
-        FXMLLoader fxmlLoader = new FXMLLoader(Game.class.getResource("hangmanScene.fxml"));
+        FXMLLoader fxmlLoader = new FXMLLoader(Game.class.getResource("GameView.fxml"));
         stage.setTitle("Hangman");
         stage.setScene(new Scene(fxmlLoader.load()));
         stage.show();
