--- src/main/resources/Game/feature_DisplayLetter.fxml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-
-<AnchorPane prefHeight="819.0" prefWidth="1024.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Game.DisplayLetter">
-    <children>
-        <Button fx:id="button_Enter" layoutX="935.0" layoutY="597.0" mnemonicParsing="false" onAction="#button" text="ENTER"  textFill="#3912a3"/>
-        <Label fx:id="wordGuess" layoutX="786.0" layoutY="566.0" prefHeight="31.0" prefWidth="204.0" text="*  GUESS WORD  *" textFill="#e41397">
-            <font>
-                <Font size="13.0" />
-            </font>
-        </Label>
-        <TextField fx:id="userInput" layoutX="786.0" layoutY="598.0" onAction="#button" />
-
-    </children>
-</AnchorPane>
