--- src/main/resources/Game/hangmanScene.fxml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import java.util.*?>
-<?import javafx.scene.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<AnchorPane xmlns="http://javafx.com/javafx"
-            xmlns:fx="http://javafx.com/fxml"
-            fx:controller="Game.ControlHangmanScene"
-            prefHeight="400.0" prefWidth="600.0">
-    <fx:include source="buttons.fxml"/>
-    <fx:include source="userInput.fxml"/>
-    <fx:include source="feature_DisplayLetter.fxml"/>
-
-
-</AnchorPane>
