--- src/main/resources/Game/userInput.fxml
@@ -1,19 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/19" xmlns:fx="http://javafx.com/fxml/1" fx:controller="Game.Word">
-   <children>
-      <Button fx:id="btn_enter" layoutX="252.0" layoutY="200.0" mnemonicParsing="false" onAction="#btn" text="Enter" />
-      <Label fx:id="wordToGuess" layoutX="236.0" layoutY="91.0" text="Label">
-         <font>
-            <Font size="36.0" />
-         </font>
-      </Label>
-      <TextField fx:id="userInput" layoutX="203.0" layoutY="153.0" onAction="#btn" />
-   </children>
-</AnchorPane>
